Create main3.c
I had to copy and paste my code because I am having the same issue with pushing my code as I did this semester during Intro to Mobile App development. My professor for that class did something to make it work, but I do not have time to fiddle around with it.